made Jernbanestasjon mogg decimal

+++ jernbane.sql
@@ -14,7 +14,7 @@ CREATE TABLE Banestrekning (
 CREATE TABLE Jernbanestasjon (
     stasjonId INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT,
     navn varchar(60) NOT NULL,
-    moh INTEGER NOT NULL
+    moh DECIMAL(6, 1) NOT NULL
 );
 
 CREATE TABLE Delstrekning (
